Move SliceViewerDataView into its own file
